Resolve conflicts

+++ Focal_app/database/migrations/2023_11_14_124316_create_business_owners_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('company_size')->nullable();
             $table->date('year_founded')->nullable();
             $table->string('responsible_job_role');
-            $table->integer('company_number');
+            $table->bigInteger('company_number');
             $table->string('website');
             $table->softDeletes();
             $table->timestamps();
